상품 페이지 : 상품 정보를 불러올때 정보 유무를 체크하다록 수정 상품 일괄 삭제 : 계정 정보 확인하도록 수정

+++ app/src/main/java/com/u/marketapp/utils/FireStoreUtils.kt
@@ -33,12 +33,16 @@ class FireStoreUtils : AppCompatActivity() {
         db.collection(activity.resources.getString(R.string.db_user))
             .document(FirebaseAuth.getInstance().currentUser!!.uid)
             .get()
-            .addOnSuccessListener { document ->
-                val user = document.toObject(UserEntity::class.java)!!
-                if (user.salesArray.size > 0) {
-                    for (sale in user.salesArray) {
-                        deleteProduct(sale)
+            .addOnSuccessListener { documentSnapshot ->
+                if (documentSnapshot.exists()) {
+                    val user = documentSnapshot.toObject(UserEntity::class.java)!!
+                    if (user.salesArray.size > 0) {
+                        for (sale in user.salesArray) {
+                            deleteProduct(sale)
+                        }
                     }
+                } else {
+                    Log.i(TAG, "유저 정보가 없음!")
                 }
             }.addOnFailureListener { e ->
                 Log.i(TAG, e.toString())
